api/requirements.txt
bert_score==0.3.13
flagembedding==1.2.7
Flask==3.1.1
numpy==2.2.6
openai==1.84.0
pandas==2.3.0
protobuf==6.31.1
python-dotenv==1.1.0
Requests==2.32.3
tiktoken==0.9.0
upstash_vector==0.8.0
